+++ grid4.h
@@ -20,7 +20,7 @@ class Grid4 {
   // 16 grid points around projection of each point:
   // first: array of index in _volume of the points
   // second: array of weights
-  std::pair<int[16], float[16]>* _projection;
+  std::pair<int[16], double[16]>* _projection;
   
   // 33 grid points around each point and itself
   Neighbours* _neighbour;
@@ -28,18 +28,18 @@ class Grid4 {
   // density
   Function _rho;
 
-  float _t;
-  float _h;
+  double _t;
+  double _h;
 
   // grid dimensions and origins (needed for index arithmetic in createProjection)
   int _wPoints;
   int _xPoints;
   int _yPoints;
   int _zPoints;
-  float _wMin;
-  float _xMin;
-  float _yMin;
-  float _zMin;
+  double _wMin;
+  double _xMin;
+  double _yMin;
+  double _zMin;
 
   public:
   // ATTENZIONE
@@ -47,9 +47,9 @@ class Grid4 {
     Interval x = {0,10},
     Interval y = {0,10},
     Interval z = {0,10},
-    float T = 40.,
-    float h = 0.1);
-  Grid4(std::string surfFile, std::string realFile, float T);
+    double T = 40.,
+    double h = 0.1);
+  Grid4(std::string surfFile, std::string realFile, double T);
   ~Grid4();
 
   void saveProjection(std::string filename = "simulation");
@@ -57,12 +57,12 @@ class Grid4 {
   void saveSurface(std::string filename = "simulation");
   void saveRho(std::string filename = "simulation");
 
-  void evolve(float dt = 0);
+  void evolve(double dt = 0);
   void project();
 
   private:
   //signed distance from the sphere
-  float phi(float w, float x, float y, float z);
+  double phi(double w, double x, double y, double z);
 
   //create the first rho value
   void createRho();
@@ -84,13 +84,13 @@ class Grid4 {
   bool loadRho(std::string filename = "simulation");
 
   //first derivative along one direction
-  float der1(int pointIndex, int direction) const;
+  double der1(int pointIndex, int direction) const;
 
   //second derivative along the same direction
-  float der2(int pointIndex, int direction) const;
+  double der2(int pointIndex, int direction) const;
 
   //second mix derivative along two different directions
-  float derij(int pointIndex, int dir1, int dir2) const;
+  double derij(int pointIndex, int dir1, int dir2) const;
 
   //logic end of int between -1 and +1
   int sgn(int val1, int val2);
